#29 refactor : OpenAi를 사용하는 ClientPort 형식을 리팩토링 한다 - 본문 요약과 코멘트 요약 기능을 담는 하나의 인터페이스를 생성한다
