--- core/src/main/java/kusitms/duduk/core/ai/dto/request/OpenAIMessages.java
@@ -1,11 +1,0 @@
-package kusitms.duduk.core.ai.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Builder
-public record OpenAIMessages (String role, String content){
-
-}
