--- core/src/main/java/kusitms/duduk/core/ai/dto/response/OpenAIResponseContent.java
@@ -1,9 +1,0 @@
-package kusitms.duduk.core.ai.dto.response;
-
-import lombok.Getter;
-
-public record OpenAIResponseContent (Message message){
-
-    public record Message (String role, String content){
-    }
-}
