--- core/src/main/java/kusitms/duduk/core/ai/dto/response/ParsedAiContentResponse.java
@@ -1,8 +1,0 @@
-package kusitms.duduk.core.ai.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record ParsedAiContentResponse(String title, String content, String keywords, String category) {
-
-}
