--- core/src/main/java/kusitms/duduk/core/ai/port/output/AiClientPort.java
@@ -1,8 +1,0 @@
-package kusitms.duduk.core.ai.port.output;
-
-import kusitms.duduk.core.ai.dto.response.OpenAIResponse;
-import kusitms.duduk.core.crawler.dto.response.CrawlingNewsResponse;
-
-public interface AiClientPort {
-    OpenAIResponse retrieveAiResponse(CrawlingNewsResponse crawlingNews);
-}
