--- core/src/main/java/kusitms/duduk/core/comment/dto/request/OpenAISummaryRequest.java
@@ -1,8 +1,0 @@
-package kusitms.duduk.core.comment.dto.request;
-
-import lombok.Builder;
-
-@Builder
-public record OpenAISummaryRequest(String comment) {
-
-}
