+++ core/src/main/java/kusitms/duduk/core/crawling/port/output/NewsCrawlingPort.java
@@ -1,0 +1,8 @@
+package kusitms.duduk.core.crawling.port.output;
+
+import java.net.MalformedURLException;
+import kusitms.duduk.core.crawling.dto.response.CrawlingNewsResponse;
+
+public interface NewsCrawlingPort {
+    public CrawlingNewsResponse crawl() throws InterruptedException, MalformedURLException;
+}
