+++ core/src/main/java/kusitms/duduk/core/openai/dto/request/OpenAiSummaryCommentRequest.java
@@ -1,0 +1,5 @@
+package kusitms.duduk.core.openai.dto.request;
+
+public record OpenAiSummaryCommentRequest(String comment) {
+
+}
