+++ core/src/main/java/kusitms/duduk/core/openai/dto/request/OpenAiSummaryNewsLetterRequest.java
@@ -1,0 +1,5 @@
+package kusitms.duduk.core.openai.dto.request;
+
+public record OpenAiSummaryNewsLetterRequest(String content) {
+
+}
