+++ core/src/main/java/kusitms/duduk/core/openai/dto/response/OpenAiSummaryNewsLetterResponse.java
@@ -1,0 +1,9 @@
+package kusitms.duduk.core.openai.dto.response;
+
+public record OpenAiSummaryNewsLetterResponse(
+    String keywords,
+    String category,
+    String content
+) {
+
+}
