--- core/src/main/java/kusitms/duduk/core/openai/port/output/OpenAISummaryClientPort.java
@@ -1,9 +1,0 @@
-package kusitms.duduk.core.openai.port.output;
-
-import kusitms.duduk.core.comment.dto.request.OpenAISummaryRequest;
-
-public interface OpenAISummaryClientPort {
-
-    String summarize(OpenAISummaryRequest request);
-
-}
