+++ core/src/main/java/kusitms/duduk/core/openai/port/output/OpenAiClientPort.java
@@ -1,0 +1,7 @@
+package kusitms.duduk.core.openai.port.output;
+
+public interface OpenAiClientPort<T, R> {
+
+    R chat(T request);
+
+}
